+++ db/migrate/20160224182015_remove_sub_section_from_sections.rb
@@ -1,0 +1,5 @@
+class RemoveSubSectionFromSections < ActiveRecord::Migration
+  def change
+    remove_column :sections, :sub_section, :boolean, null: false, default: false
+  end
+end